+++ packages/preferences/src/browser/preferences-monaco-contribution.ts
@@ -23,8 +23,5 @@ monaco.languages.register({
     ],
     'filenames': [
         'settings.json'
-    ],
-    'extensions': [
-        '.theia-workspace'
     ]
 });