+++ packages/workspace/src/node/workspace-backend-module.ts
@@ -16,7 +16,7 @@
 
 import { ContainerModule } from '@theia/core/shared/inversify';
 import { ConnectionHandler, RpcConnectionHandler } from '@theia/core/lib/common';
-import { WorkspaceServer, workspacePath, CommonWorkspaceUtils } from '../common';
+import { WorkspaceServer, workspacePath, UntitledWorkspaceService, WorkspaceFileService } from '../common';
 import { DefaultWorkspaceServer, WorkspaceCliContribution } from './default-workspace-server';
 import { CliContribution } from '@theia/core/lib/node/cli';
 import { BackendApplicationContribution } from '@theia/core/lib/node';
@@ -27,7 +27,8 @@ export default new ContainerModule(bind => {
     bind(DefaultWorkspaceServer).toSelf().inSingletonScope();
     bind(WorkspaceServer).toService(DefaultWorkspaceServer);
     bind(BackendApplicationContribution).toService(WorkspaceServer);
-    bind(CommonWorkspaceUtils).toSelf().inSingletonScope();
+    bind(UntitledWorkspaceService).toSelf().inSingletonScope();
+    bind(WorkspaceFileService).toSelf().inSingletonScope();
 
     bind(ConnectionHandler).toDynamicValue(ctx =>
         new RpcConnectionHandler(workspacePath, () =>